// src/lib/template-types-generated.ts
export type MissionDataName =
  | "Abductions"
  | "Advise"
  | "AppeaseWin"
  | "Assassinate"
  | "AssaultAlienAsset"
  | "AssumeControl"
  | "BuildFacility"
  | "Contact"
  | "ControlSpaceAsset"
  | "CooperateWin"
  | "Coup"
  | "Crackdown"
  | "DefendInterests"
  | "Deorbit"
  | "DestroyWin"
  | "Detain"
  | "DetectCouncilActivity"
  | "DominateNation"
  | "EnthrallElites"
  | "EnthrallOrg"
  | "EnthrallPublic"
  | "EnthrallUnalignedElites"
  | "EscapeWin"
  | "ExploitWin"
  | "Extract"
  | "GainInfluence"
  | "GoToGround"
  | "HostileTakeover"
  | "Inspire"
  | "InvestigateAlienActivity"
  | "InvestigateCouncilor"
  | "Orbit"
  | "PassTechnology"
  | "Propaganda"
  | "Protect"
  | "Purge"
  | "ResistWin"
  | "SabotageFacilities"
  | "SabotageHabModule"
  | "SabotageProject"
  | "SeizeSpaceAsset"
  | "SetNationalPolicy"
  | "Stabilize"
  | "StealProject"
  | "SubmitWin"
  | "TerrorizeRegion"
  | "Transfer"
  | "Turn"
  | "Unrest"
  | "Xenoform";
// ...
export type TechCategory =
  | "Energy"
  | "InformationScience"
  | "LifeScience"
  | "Materials"
  | "MilitaryScience"
  | "SocialScience"
  | "SpaceScience";